[other] Maintain the doc, development settings.
Rename authentication to authorization.
