+++ src/main/java/top/horizonask/hoawiki/HoawikiApplication.java
@@ -5,7 +5,7 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.mybatis.spring.annotation.MapperScan;
 
 @SpringBootApplication()
-@MapperScan("top.horizonask.hoawiki.authentication.mapper")
+@MapperScan("top.horizonask.hoawiki.authorization.mapper")
 public class HoawikiApplication {
     public static void main(String[] args) {
         SpringApplication.run(HoawikiApplication.class, args);
