--- src/main/java/top/horizonask/hoawiki/authentication/entity/UsersRole.java
@@ -1,19 +1,0 @@
-package top.horizonask.hoawiki.authentication.entity;
-
-import com.baomidou.mybatisplus.annotation.TableField;
-import com.baomidou.mybatisplus.annotation.TableName;
-import com.baomidou.mybatisplus.extension.activerecord.Model;
-import lombok.Data;
-import lombok.EqualsAndHashCode;
-
-@EqualsAndHashCode(callSuper = true)
-@Data
-@TableName(value = "users_role")
-public class UsersRole extends Model<UsersRole> {
-
-    @TableField(value = "user_id")
-    private Long userId;
-
-    @TableField(value = "role_id")
-    private Long roleId;
-}
