--- src/main/java/top/horizonask/hoawiki/authentication/mapper/RolesPermissionMapper.java
@@ -1,10 +1,0 @@
-package top.horizonask.hoawiki.authentication.mapper;
-
-import top.horizonask.hoawiki.authentication.entity.RolesPermission;
-import com.baomidou.mybatisplus.core.mapper.BaseMapper;
-import org.apache.ibatis.annotations.Mapper;
-
-@Mapper
-public interface RolesPermissionMapper extends BaseMapper<RolesPermission> {
-
-}
