--- src/main/java/top/horizonask/hoawiki/authentication/mapper/UsersRoleMapper.java
@@ -1,10 +1,0 @@
-package top.horizonask.hoawiki.authentication.mapper;
-
-import top.horizonask.hoawiki.authentication.entity.UsersRole;
-import com.baomidou.mybatisplus.core.mapper.BaseMapper;
-import org.apache.ibatis.annotations.Mapper;
-
-@Mapper
-public interface UsersRoleMapper extends BaseMapper<UsersRole> {
-
-}
