--- src/main/java/top/horizonask/hoawiki/authentication/service/UserService.java
@@ -1,7 +1,0 @@
-package top.horizonask.hoawiki.authentication.service;
-
-import com.baomidou.mybatisplus.extension.service.IService;
-import top.horizonask.hoawiki.authentication.entity.User;
-
-public interface UserService extends IService<User> {
-}
