--- src/main/java/top/horizonask/hoawiki/authentication/service/UserServiceImpl.java
@@ -1,14 +1,0 @@
-package top.horizonask.hoawiki.authentication.service;
-
-import top.horizonask.hoawiki.authentication.mapper.UserMapper;
-import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
-import top.horizonask.hoawiki.authentication.entity.User;
-
-/**
- * @description:
- * @author: Yanbo Han
- * @time: 2021/12/31 0:18
- */
-
-public class UserServiceImpl extends ServiceImpl<UserMapper, User> {
-}
