+++ src/main/java/top/horizonask/hoawiki/authorization/entity/UsersRole.java
@@ -1,0 +1,19 @@
+package top.horizonask.hoawiki.authorization.entity;
+
+import com.baomidou.mybatisplus.annotation.TableField;
+import com.baomidou.mybatisplus.annotation.TableName;
+import com.baomidou.mybatisplus.extension.activerecord.Model;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+
+@EqualsAndHashCode(callSuper = true)
+@Data
+@TableName(value = "users_role")
+public class UsersRole extends Model<UsersRole> {
+
+    @TableField(value = "user_id")
+    private Long userId;
+
+    @TableField(value = "role_id")
+    private Long roleId;
+}
