+++ src/main/java/top/horizonask/hoawiki/authorization/handler/TimeObjectHandler.java
@@ -1,0 +1,27 @@
+package top.horizonask.hoawiki.authorization.handler;
+
+import com.baomidou.mybatisplus.core.handlers.MetaObjectHandler;
+import org.apache.ibatis.reflection.MetaObject;
+import org.springframework.stereotype.Component;
+
+import java.time.LocalDateTime;
+
+/**
+ * @description:
+ * @author: Yanbo Han
+ * @time: 2022/1/1 18:24
+ */
+
+@Component
+public class TimeObjectHandler implements MetaObjectHandler {
+    @Override
+    public void insertFill(MetaObject metaObject){
+        this.strictInsertFill(metaObject, "createTime", LocalDateTime.class, LocalDateTime.now());
+        this.strictInsertFill(metaObject, "updateTime", LocalDateTime.class, LocalDateTime.now());
+    }
+
+    @Override
+    public void updateFill(MetaObject metaObject){
+        this.strictInsertFill(metaObject, "updateTime", LocalDateTime.class, LocalDateTime.now());
+    }
+}
