+++ src/main/java/top/horizonask/hoawiki/authorization/mapper/RolesPermissionMapper.java
@@ -1,0 +1,10 @@
+package top.horizonask.hoawiki.authorization.mapper;
+
+import top.horizonask.hoawiki.authorization.entity.RolesPermission;
+import com.baomidou.mybatisplus.core.mapper.BaseMapper;
+import org.apache.ibatis.annotations.Mapper;
+
+@Mapper
+public interface RolesPermissionMapper extends BaseMapper<RolesPermission> {
+
+}
