+++ src/main/java/top/horizonask/hoawiki/authorization/mapper/UsersRoleMapper.java
@@ -1,0 +1,10 @@
+package top.horizonask.hoawiki.authorization.mapper;
+
+import top.horizonask.hoawiki.authorization.entity.UsersRole;
+import com.baomidou.mybatisplus.core.mapper.BaseMapper;
+import org.apache.ibatis.annotations.Mapper;
+
+@Mapper
+public interface UsersRoleMapper extends BaseMapper<UsersRole> {
+
+}
