+++ src/main/java/top/horizonask/hoawiki/authorization/service/UserService.java
@@ -1,0 +1,7 @@
+package top.horizonask.hoawiki.authorization.service;
+
+import com.baomidou.mybatisplus.extension.service.IService;
+import top.horizonask.hoawiki.authorization.entity.User;
+
+public interface UserService extends IService<User> {
+}
