+++ src/main/java/top/horizonask/hoawiki/authorization/service/UserServiceImpl.java
@@ -1,0 +1,14 @@
+package top.horizonask.hoawiki.authorization.service;
+
+import top.horizonask.hoawiki.authorization.mapper.UserMapper;
+import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
+import top.horizonask.hoawiki.authorization.entity.User;
+
+/**
+ * @description:
+ * @author: Yanbo Han
+ * @time: 2021/12/31 0:18
+ */
+
+public class UserServiceImpl extends ServiceImpl<UserMapper, User> {
+}
